503: 2 solutions: jumping, monotonic stack.
The "jumping" solution "jumps" using already calculated larger indices.
Although its time complexity is higher, it runs faster than the stack solution.

The "monotonic stack" solution maintains a stack that only stores monotonically increasing larger elements.
We maintain the stack so that the stack top is the smallest element that is larger than the current element and is the answer.
If we encounter a smaller (or equal) element on the stack top, we pop it to try to find a bigger one.
If the stack is empty, the current element is the largest.